backend/requirements.txt
Flask
Flask-Mail
python-dotenv
